backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
requests>=2.31.0
langfuse>=2.20.0
pillow>=10.2.0
PyPDF2>=3.0.1
pdf2image>=1.17.0
python-dotenv>=1.0.0
groq>=0.4.0
openai>=1.0.0
sqlalchemy>=2.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.1.0
